LayoutTests/fast/css-grid-layout/grid-content-alignment-and-self-alignment.html

<!DOCTYPE html>
<html>
<head>
<link href="resources/grid.css" rel="stylesheet">
<script src="../../resources/check-layout.js"></script>
<style>
body {
    margin: 0px;
}

.grid {
    -webkit-grid-auto-columns: 20px;
    -webkit-grid-auto-rows: 40px;
    -webkit-grid-template-areas: "a a b"
                                 "c d b";
    position: relative;
    width: 300px;
    height: 200px;
}
.a {
    -webkit-grid-area: a;
    background-color: blue;
}
.b {
    -webkit-grid-area: b;
    background-color: lime;
}
.c {
    -webkit-grid-area: c;
    background-color: purple;
}
.d {
    -webkit-grid-area: d;
    background-color: orange;
}
.stretchedGrid {
    -webkit-grid-auto-columns: minmax(20px, auto);
    -webkit-grid-auto-rows: minmax(40px, auto);
}

.spaceBetween {
    justify-content: space-between;
    align-content: space-between;
}

.spaceAround {
    justify-content: space-around;
    align-content: space-around;
}

.spaceEvenly {
    justify-content: space-evenly;
    align-content: space-evenly;
}

.stretch {
    justify-content: stretch;
    align-content: stretch;
}

.cell {
    width: 20px;
    height: 40px;
}
.justifyCenter {
    justify-self: center;
}
.alignCenter {
    align-self: center;
}
.justifyEnd {
    justify-self: end;
}
.alignEnd {
    align-self: end;
}
</style>
</head>
<body onload="checkLayout('.grid')">

<p>This test checks that Content Distribution alignment works fine in combination with Self Alignmet and items span more than one track.</p>

<div style="position: relative">
    <p>direction: LTR | distribution: 'space-between' | self-alignment: center</p>
    <div class="grid spaceBetween" data-expected-width="300" data-expected-height="200">
        <div class="a cell justifyCenter" data-offset-x="70" data-offset-y="0" data-expected-width="20" data-expected-height="40"></div>
        <div class="b cell alignCenter" data-offset-x="280" data-offset-y="80" data-expected-width="20" data-expected-height="40"></div>
        <div class="c" data-offset-x="0" data-offset-y="160" data-expected-width="20" data-expected-height="40"></div>
        <div class="d" data-offset-x="140" data-offset-y="160" data-expected-width="20" data-expected-height="40"></div>
    </div>
</div>

<div style="position: relative">
    <p>direction: LTR | distribution: 'space-between' | self-alignment: end</p>
    <div class="grid spaceBetween" data-expected-width="300" data-expected-height="200">
        <div class="a cell justifyEnd" data-offset-x="140" data-offset-y="0" data-expected-width="20" data-expected-height="40"></div>
        <div class="b cell alignEnd" data-offset-x="280" data-offset-y="160" data-expected-width="20" data-expected-height="40"></div>
        <div class="c" data-offset-x="0" data-offset-y="160" data-expected-width="20" data-expected-height="40"></div>
        <div class="d" data-offset-x="140" data-offset-y="160" data-expected-width="20" data-expected-height="40"></div>
    </div>
</div>

<div style="position: relative">
    <p>direction: LTR | distribution: 'space-around' | self-alignment: center</p>
    <div class="grid spaceAround" data-expected-width="300" data-expected-height="200">
        <div class="a cell justifyCenter" data-offset-x="90" data-offset-y="30" data-expected-width="20" data-expected-height="40"></div>
        <div class="b cell alignCenter" data-offset-x="240" data-offset-y="80" data-expected-width="20" data-expected-height="40"></div>
        <div class="c" data-offset-x="40" data-offset-y="130" data-expected-width="20" data-expected-height="40"></div>
        <div class="d" data-offset-x="140" data-offset-y="130" data-expected-width="20" data-expected-height="40"></div>
    </div>
</div>

<div style="position: relative">
    <p>direction: LTR | distribution: 'space-around' | self-alignment: end</p>
    <div class="grid spaceAround" data-expected-width="300" data-expected-height="200">
        <div class="a cell justifyEnd" data-offset-x="140" data-offset-y="30" data-expected-width="20" data-expected-height="40"></div>
        <div class="b cell alignEnd" data-offset-x="240" data-offset-y="130" data-expected-width="20" data-expected-height="40"></div>
        <div class="c" data-offset-x="40" data-offset-y="130" data-expected-width="20" data-expected-height="40"></div>
        <div class="d" data-offset-x="140" data-offset-y="130" data-expected-width="20" data-expected-height="40"></div>
    </div>
</div>

<div style="position: relative">
    <p>direction: LTR | distribution: 'space-evenly' | self-alignment: center</p>
    <div class="grid spaceEvenly" data-expected-width="300" data-expected-height="200">
        <div class="a cell justifyCenter" data-offset-x="100" data-offset-y="40" data-expected-width="20" data-expected-height="40"></div>
        <div class="b cell alignCenter" data-offset-x="220" data-offset-y="80" data-expected-width="20" data-expected-height="40"></div>
        <div class="c" data-offset-x="60" data-offset-y="120" data-expected-width="20" data-expected-height="40"></div>
        <div class="d" data-offset-x="140" data-offset-y="120" data-expected-width="20" data-expected-height="40"></div>
    </div>
</div>

<div style="position: relative">
    <p>direction: LTR | distribution: 'space-evenly' | self-alignment: end</p>
    <div class="grid spaceEvenly" data-expected-width="300" data-expected-height="200">
        <div class="a cell justifyEnd" data-offset-x="140" data-offset-y="40" data-expected-width="20" data-expected-height="40"></div>
        <div class="b cell alignEnd" data-offset-x="220" data-offset-y="120" data-expected-width="20" data-expected-height="40"></div>
        <div class="c" data-offset-x="60" data-offset-y="120" data-expected-width="20" data-expected-height="40"></div>
        <div class="d" data-offset-x="140" data-offset-y="120" data-expected-width="20" data-expected-height="40"></div>
    </div>
</div>

<div style="position: relative">
    <p>direction: LTR | distribution: 'stretch' | self-alignment: center</p>
    <div class="grid stretchedGrid stretch" data-expected-width="300" data-expected-height="200">
        <div class="a cell justifyCenter alignCenter" data-offset-x="90" data-offset-y="30" data-expected-width="20" data-expected-height="40"></div>
        <div class="b cell justifyCenter alignCenter" data-offset-x="240" data-offset-y="80" data-expected-width="20" data-expected-height="40"></div>
        <div class="c" data-offset-x="0" data-offset-y="100" data-expected-width="100" data-expected-height="100"></div>
        <div class="d" data-offset-x="100" data-offset-y="100" data-expected-width="100" data-expected-height="100"></div>
    </div>
</div>

<div style="position: relative">
    <p>direction: LTR | distribution: 'stretch' | self-alignment: end</p>
    <div class="grid stretchedGrid stretch" data-expected-width="300" data-expected-height="200">
        <div class="a cell justifyEnd alignEnd" data-offset-x="180" data-offset-y="60" data-expected-width="20" data-expected-height="40"></div>
        <div class="b cell justifyEnd alignEnd" data-offset-x="280" data-offset-y="160" data-expected-width="20" data-expected-height="40"></div>
        <div class="c" data-offset-x="0" data-offset-y="100" data-expected-width="100" data-expected-height="100"></div>
        <div class="d" data-offset-x="100" data-offset-y="100" data-expected-width="100" data-expected-height="100"></div>
    </div>
</div>

<!-- RTL direction. -->
<div style="position: relative">
    <p>direction: RTL | distribution: 'space-between' | self-alignment: center</p>
    <div class="grid spaceBetween directionRTL" data-expected-width="300" data-expected-height="200">
        <div class="a cell justifyCenter" data-offset-x="210" data-offset-y="0" data-expected-width="20" data-expected-height="40"></div>
        <div class="b cell alignCenter" data-offset-x="0" data-offset-y="80" data-expected-width="20" data-expected-height="40"></div>
        <div class="c" data-offset-x="280" data-offset-y="160" data-expected-width="20" data-expected-height="40"></div>
        <div class="d" data-offset-x="140" data-offset-y="160" data-expected-width="20" data-expected-height="40"></div>
    </div>
</div>

<div style="position: relative">
    <p>direction: RTL | distribution: 'space-between' | self-alignment: end</p>
    <div class="grid spaceBetween directionRTL" data-expected-width="300" data-expected-height="200">
        <div class="a cell justifyEnd" data-offset-x="140" data-offset-y="0" data-expected-width="20" data-expected-height="40"></div>
        <div class="b cell alignEnd" data-offset-x="0" data-offset-y="160" data-expected-width="20" data-expected-height="40"></div>
        <div class="c" data-offset-x="280" data-offset-y="160" data-expected-width="20" data-expected-height="40"></div>
        <div class="d" data-offset-x="140" data-offset-y="160" data-expected-width="20" data-expected-height="40"></div>
    </div>
</div>

<div style="position: relative">
    <p>direction: RTL | distribution: 'space-around' | self-alignment: center</p>
    <div class="grid spaceAround directionRTL" data-expected-width="300" data-expected-height="200">
        <div class="a cell justifyCenter" data-offset-x="190" data-offset-y="30" data-expected-width="20" data-expected-height="40"></div>
        <div class="b cell alignCenter" data-offset-x="40" data-offset-y="80" data-expected-width="20" data-expected-height="40"></div>
        <div class="c" data-offset-x="240" data-offset-y="130" data-expected-width="20" data-expected-height="40"></div>
        <div class="d" data-offset-x="140" data-offset-y="130" data-expected-width="20" data-expected-height="40"></div>
    </div>
</div>

<div style="position: relative">
    <p>direction: RTL | distribution: 'space-around' | self-alignment: end</p>
    <div class="grid spaceAround directionRTL" data-expected-width="300" data-expected-height="200">
        <div class="a cell justifyEnd" data-offset-x="140" data-offset-y="30" data-expected-width="20" data-expected-height="40"></div>
        <div class="b cell alignEnd" data-offset-x="40" data-offset-y="130" data-expected-width="20" data-expected-height="40"></div>
        <div class="c" data-offset-x="240" data-offset-y="130" data-expected-width="20" data-expected-height="40"></div>
        <div class="d" data-offset-x="140" data-offset-y="130" data-expected-width="20" data-expected-height="40"></div>
    </div>
</div>

<div style="position: relative">
    <p>direction: RTL | distribution: 'space-evenly' | self-alignment: center</p>
    <div class="grid spaceEvenly directionRTL" data-expected-width="300" data-expected-height="200">
        <div class="a cell justifyCenter" data-offset-x="180" data-offset-y="40" data-expected-width="20" data-expected-height="40"></div>
        <div class="b cell alignCenter" data-offset-x="60" data-offset-y="80" data-expected-width="20" data-expected-height="40"></div>
        <div class="c" data-offset-x="220" data-offset-y="120" data-expected-width="20" data-expected-height="40"></div>
        <div class="d" data-offset-x="140" data-offset-y="120" data-expected-width="20" data-expected-height="40"></div>
    </div>
</div>

<div style="position: relative">
    <p>direction: RTL | distribution: 'space-evenly' | self-alignment: end</p>
    <div class="grid spaceEvenly directionRTL" data-expected-width="300" data-expected-height="200">
        <div class="a cell justifyEnd" data-offset-x="140" data-offset-y="40" data-expected-width="20" data-expected-height="40"></div>
        <div class="b cell alignEnd" data-offset-x="60" data-offset-y="120" data-expected-width="20" data-expected-height="40"></div>
        <div class="c" data-offset-x="220" data-offset-y="120" data-expected-width="20" data-expected-height="40"></div>
        <div class="d" data-offset-x="140" data-offset-y="120" data-expected-width="20" data-expected-height="40"></div>
    </div>
</div>

<div style="position: relative">
    <p>direction: RTL | distribution: 'stretch' | self-alignment: center</p>
    <div class="grid stretchedGrid stretch directionRTL" data-expected-width="300" data-expected-height="200">
        <div class="a cell justifyCenter alignCenter" data-offset-x="190" data-offset-y="30" data-expected-width="20" data-expected-height="40"></div>
        <div class="b cell justifyCenter alignCenter" data-offset-x="40" data-offset-y="80" data-expected-width="20" data-expected-height="40"></div>
        <div class="c" data-offset-x="200" data-offset-y="100" data-expected-width="100" data-expected-height="100"></div>
        <div class="d" data-offset-x="100" data-offset-y="100" data-expected-width="100" data-expected-height="100"></div>
    </div>
</div>

<div style="position: relative">
    <p>direction: RTL | distribution: 'stretch' | self-alignment: end</p>
    <div class="grid stretchedGrid stretch directionRTL" data-expected-width="300" data-expected-height="200">
        <div class="a cell justifyEnd alignEnd" data-offset-x="100" data-offset-y="60" data-expected-width="20" data-expected-height="40"></div>
        <div class="b cell justifyEnd alignEnd" data-offset-x="0" data-offset-y="160" data-expected-width="20" data-expected-height="40"></div>
        <div class="c" data-offset-x="200" data-offset-y="100" data-expected-width="100" data-expected-height="100"></div>
        <div class="d" data-offset-x="100" data-offset-y="100" data-expected-width="100" data-expected-height="100"></div>
    </div>
</div>

</body>
</html>
