Safari-609.1.11.1/LayoutTests/fast/css-grid-layout/grid-content-alignment-and-self-alignment-expected.txt

This test checks that Content Distribution alignment works fine in combination with Self Alignmet and items in just one cell.

direction: LTR | distribution: 'space-between' | self-alignment: center

X
X
X
X
PASS
direction: LTR | distribution: 'space-between' | self-alignment: end

X
X
X
X
PASS
direction: LTR | distribution: 'space-around' | self-alignment: center

X
X
X
X
PASS
direction: LTR | distribution: 'space-around' | self-alignment: end

X
X
X
X
PASS
direction: LTR | distribution: 'space-evenly' | self-alignment: center

X
X
X
X
PASS
direction: LTR | distribution: 'space-evenly' | self-alignment: end

X
X
X
X
PASS
direction: LTR | grid-gap: 10px 20px | distribution: 'space-between' | self-alignment: center

X
X
X
X
PASS
direction: LTR | grid-gap: 10px 20px | distribution: 'space-between' | self-alignment: end

X
X
X
X
PASS
direction: LTR | grid-gap: 10px 20px | distribution: 'space-around' | self-alignment: center

X
X
X
X
PASS
direction: LTR | grid-gap: 10px 20px | distribution: 'space-around' | self-alignment: end

X
X
X
X
PASS
direction: LTR | grid-gap: 10px 20px | distribution: 'space-evenly' | self-alignment: center

X
X
X
X
PASS
direction: LTR | grid-gap: 10px 20px | distribution: 'space-evenly' | self-alignment: end

X
X
X
X
PASS
direction: RTL | distribution: 'space-between' | self-alignment: center

X
X
X
X
PASS
direction: RTL | distribution: 'space-between' | self-alignment: end

X
X
X
X
PASS
direction: RTL | distribution: 'space-around' | self-alignment: center

X
X
X
X
PASS
direction: RTL | distribution: 'space-around' | self-alignment: end

X
X
X
X
PASS
direction: RTL | distribution: 'space-evenly' | self-alignment: center

X
X
X
X
PASS
direction: RTL | distribution: 'space-evenly' | self-alignment: end

X
X
X
X
PASS
direction: RTL | grid-gap: 10px 20px | distribution: 'space-between' | self-alignment: center

X
X
X
X
PASS
direction: RTL | grid-gap: 10px 20px | distribution: 'space-between' | self-alignment: end

X
X
X
X
PASS
direction: RTL | grid-gap: 10px 20px | distribution: 'space-around' | self-alignment: center

X
X
X
X
PASS
direction: RTL | grid-gap: 10px 20px | distribution: 'space-around' | self-alignment: end

X
X
X
X
PASS
direction: RTL | grid-gap: 10px 20px | distribution: 'space-evenly' | self-alignment: center

X
X
X
X
PASS
direction: RTL | grid-gap: 10px 20px | distribution: 'space-evenly' | self-alignment: end

X
X
X
X
PASS
